[EDIT] Has Key, Trim 구현

+++ controllers/cart.js
@@ -7,7 +7,6 @@ const getUserCart = async (req, res) => {
     const result = await cartService.getUserCart(
       user_id //, token
     );
-    console.log(result);
     res.status(200).json(result);
   } catch (error) {
     console.log(error);
@@ -17,6 +16,29 @@ const getUserCart = async (req, res) => {
 
 const addCart = async (req, res) => {
   //const token = req.get('authorization');
+  const hasKey = {
+    user_id: false,
+    product_id: false,
+    size_id: false,
+    count: false,
+  };
+  const requireKey = Object.keys(hasKey);
+
+  Object.entries(req.body).forEach(keyValue => {
+    const [key, value] = keyValue;
+    if (requireKey.includes(key) && value) {
+      hasKey[key] = true;
+    }
+  });
+  const hasKeyArray = Object.entries(hasKey);
+  for (let i = 0; i < hasKeyArray.length; i++) {
+    const [key, value] = hasKeyArray[i];
+    if (!value) {
+      res.status(400).json({ message: `please enter ${key}` });
+      return;
+    }
+  }
+
   const { user_id, product_id, size_id, count } = req.body;
   try {
     const result = await cartService.addCart(
@@ -26,7 +48,6 @@ const addCart = async (req, res) => {
       size_id,
       count
     );
-    console.log(result);
     res.status(200).json({ message: 'Item Added' });
   } catch (error) {
     console.log(error);
@@ -36,11 +57,14 @@ const addCart = async (req, res) => {
 
 const deleteItem = async (req, res) => {
   const { cart_id } = req.query;
+  if (!cart_id) {
+    res.status(400).json({ message: `please enter cart_id` });
+    return;
+  }
   try {
     const result = await cartService.deleteItem(
       cart_id //, token
     );
-    console.log(result);
     res.status(200).json(result);
   } catch (error) {
     console.log(error);
@@ -50,11 +74,14 @@ const deleteItem = async (req, res) => {
 
 const deleteAllItem = async (req, res) => {
   const { user_id } = req.query;
+  if (!user_id) {
+    res.status(400).json({ message: `please enter user_id` });
+    return;
+  }
   try {
     const result = await cartService.deleteAllItem(
       user_id //, token
     );
-    console.log(result);
     res.status(200).json({ massege: 'All Item Removed' });
   } catch (error) {
     console.log(error);
